refactor: Complete dependency collector architecture refactoring
- Separate AST node extraction and dependency resolution (#66)
- Rename models for better clarity (UsageNode → Usage)
- Remove unnecessary dependency_collector layer
- Modularize models.rs into individual files
- Implement external data flow pattern for better separation of concerns
