+++ core/src/collectors/common/mod.rs
@@ -1,2 +1,3 @@
 pub mod definition_collectors;
-pub mod dependency_collectors;
+pub mod dependency_resolver;
+pub mod usage_node_collector;
