+++ core/src/collectors/rust/mod.rs
@@ -1,2 +1,3 @@
 pub mod rust_definition_collector;
-pub mod rust_dependency_collector;
+pub mod rust_dependency_resolver;
+pub mod rust_usage_node_collector;
