+++ core/src/collectors/typescript/mod.rs
@@ -1,2 +1,3 @@
 pub mod typescript_definition_collector;
-pub mod typescript_dependency_collector;
+pub mod typescript_dependency_resolver;
+pub mod typescript_usage_node_collector;
