+++ core/src/models/dependency.rs
@@ -1,0 +1,22 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
+pub enum DependencyType {
+    FunctionCall,
+    VariableUse,
+    Import,
+    StructFieldAccess,
+    TypeReference,
+    ModuleReference,
+    MacroInvocation,
+    Other(String),
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct Dependency {
+    pub source_line: usize,
+    pub target_line: usize,
+    pub symbol: String,
+    pub dependency_type: DependencyType,
+    pub context: Option<String>,
+}
