+++ core/src/models/intermediate_representation.rs
@@ -1,0 +1,42 @@
+use super::{Definition, Dependency};
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct AnalysisMetadata {
+    pub language: String,
+    pub total_lines: usize,
+    pub analysis_timestamp: String,
+    pub lintric_version: String,
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct IntermediateRepresentation {
+    pub file_path: String,
+    pub definitions: Vec<Definition>,
+    pub dependencies: Vec<Dependency>,
+    pub analysis_metadata: AnalysisMetadata,
+}
+
+impl IntermediateRepresentation {
+    pub fn new(
+        file_path: String,
+        definitions: Vec<Definition>,
+        dependencies: Vec<Dependency>,
+        language: String,
+        total_lines: usize,
+    ) -> Self {
+        let analysis_metadata = AnalysisMetadata {
+            language,
+            total_lines,
+            analysis_timestamp: "now".to_string(), // Placeholder for timestamp
+            lintric_version: env!("CARGO_PKG_VERSION").to_string(),
+        };
+
+        IntermediateRepresentation {
+            file_path,
+            definitions,
+            dependencies,
+            analysis_metadata,
+        }
+    }
+}
