+++ core/src/models/usage.rs
@@ -1,0 +1,17 @@
+use tree_sitter::Node;
+
+#[derive(Debug, Clone, PartialEq)]
+pub enum UsageKind {
+    Identifier,
+    CallExpression,
+    FieldExpression,
+    StructExpression,
+    Metavariable,
+}
+
+#[derive(Debug, Clone)]
+pub struct Usage<'a> {
+    pub node: Node<'a>,
+    pub kind: UsageKind,
+    pub scope: Option<String>,
+}
